health/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.0
django-debug-toolbar==6.0.0
django-extensions==4.1
django-filter==25.1
django-lifecycle==1.2.4
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Faker==37.5.3
idna==3.10
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
redis==6.4.0
requests==2.32.4
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
gunicorn==23.0.0
boto3==1.38.20
drf-spectacular==0.28.0
django-ckeditor-5==0.2.17
django-cors-headers==4.9.0
cryptography==46.0.3
dj-rest-auth==7.0.1
langchain
langchain-ollama
langchain-core
langchain-community
